+++ src/containers/Form/OutboundIdentifierLists/layout.js
@@ -12,16 +12,28 @@ import React from 'react';
 import PropTypes from 'prop-types';
 import { InputField } from 'cx-ui-components';
 
-export default function OutboundIdentifierListsForm({ handleSubmit, isSaving, inherited, key }) {
+export default function OutboundIdentifierListsForm({
+  handleSubmit,
+  isSaving,
+  inherited,
+  userHasUpdatePermission,
+  key
+}) {
   return (
     <form onSubmit={handleSubmit} key={key}>
-      <InputField name="name" label="Name *" componentType="input" inputType="text" disabled={isSaving || inherited} />
+      <InputField
+        name="name"
+        label="Name *"
+        componentType="input"
+        inputType="text"
+        disabled={isSaving || inherited || !userHasUpdatePermission}
+      />
       <InputField
         name="description"
         label="Description"
         componentType="textarea"
         inputType="text"
-        disabled={isSaving || inherited}
+        disabled={isSaving || inherited || !userHasUpdatePermission}
       />
     </form>
   );
@@ -31,5 +43,6 @@ OutboundIdentifierListsForm.propTypes = {
   key: PropTypes.string,
   handleSubmit: PropTypes.func.isRequired,
   isSaving: PropTypes.bool,
-  inherited: PropTypes.bool
+  inherited: PropTypes.bool,
+  userHasUpdatePermission: PropTypes.bool
 };