+++ src/redux/modules/entities/__tests__/listItemSelectors.test.js
@@ -2,7 +2,7 @@
  * Copyright © 2015-2018 Serenova, LLC. All rights reserved.
  */
 
-import { fromJS, List } from 'immutable';
+import { fromJS } from 'immutable';
 import { getCurrentEntity, getSelectedEntity, getCurrentEntityStore } from '../selectors';
 import {
   listMemberIds,
@@ -21,6 +21,7 @@ import {
 import { EntityMetaData, entitiesMetaData } from '../metaData';
 
 entitiesMetaData.mockEntity = new EntityMetaData('mockEntity');
+entitiesMetaData.mockEntity.dependentEntity = 'mockDependentEntity';
 
 const initialState = fromJS({
   Entities: {
@@ -30,13 +31,7 @@ const initialState = fromJS({
           id: '0000',
           name: 'mockName',
           active: true,
-          mockDependentEntity: [
-            {
-              id: '0001',
-              name: 'mockName',
-              active: true
-            }
-          ]
+          mockDependentEntity: ['0001']
         }
       ]
     },
@@ -56,13 +51,7 @@ const mockSelectedEntity = fromJS({
   id: '0000',
   name: 'mockName',
   active: true,
-  mockDependentEntity: [
-    {
-      id: '0001',
-      name: 'mockName',
-      active: true
-    }
-  ]
+  mockDependentEntity: ['0001']
 });
 
 const mockCurrentEntityStore = fromJS({
@@ -71,13 +60,7 @@ const mockCurrentEntityStore = fromJS({
       id: '0000',
       name: 'mockName',
       active: true,
-      mockDependentEntity: [
-        {
-          id: '0001',
-          name: 'mockName',
-          active: true
-        }
-      ]
+      mockDependentEntity: ['0001']
     }
   ],
   selectedEntityId: '0000'
@@ -90,49 +73,231 @@ getCurrentEntityStore.mockImplementation(() => mockCurrentEntityStore);
 
 describe('listMemberIds', () => {
   it('should get dependentEntity, then return it as javascript collection', () => {
-    expect(listMemberIds(initialState)).toEqual([]);
+    expect(listMemberIds(initialState)).toEqual(['0001']);
   });
 });
 
 describe('listMemberObjects', () => {
   it('should get all current entity members, then return them as javascript collection', () => {
-    expect(listMemberObjects(initialState)).toEqual(new List([]));
+    expect(listMemberObjects(initialState)).toEqual(fromJS([{ id: '0001', name: 'mockName', active: true }]));
   });
 });
 
 describe('getDependantEntityTableItems', () => {
   it('should get dependent members of current entity, then return them as javascript collection', () => {
-    expect(getDependantEntityTableItems(initialState)).toEqual([]);
+    expect(getDependantEntityTableItems(initialState)).toEqual([{ active: true, id: '0001', name: 'mockName' }]);
+  });
+  it('should get empty array since dependentEntity has no entries in state', () => {
+    expect(
+      getDependantEntityTableItems(
+        fromJS({
+          Entities: {
+            mockEntity: {
+              data: [
+                {
+                  id: '0000',
+                  mockDependentEntity: ['0001']
+                }
+              ]
+            },
+            mockDependentEntity: {
+              data: undefined
+            }
+          }
+        })
+      )
+    ).toEqual([]);
   });
 });
 
 describe('getSidePanelTableItems', () => {
   it('should get all current entity members for side panel use', () => {
-    expect(getSidePanelTableItems(initialState)).toEqual(undefined);
+    expect(getSidePanelTableItems(initialState, 'mockDependentEntity')).toEqual([
+      { active: true, id: '0001', name: 'mockName' }
+    ]);
+  });
+  it('should get empty array since dependentEntity has no entries in state', () => {
+    expect(
+      getSidePanelTableItems(
+        fromJS({
+          Entities: {
+            mockEntity: {
+              data: [
+                {
+                  id: '0000',
+                  name: 'mockName',
+                  active: true,
+                  mockDependentEntity: ['0001']
+                }
+              ]
+            },
+            mockDependentEntity: {
+              data: undefined
+            }
+          }
+        }),
+        'mockDependentEntity'
+      )
+    ).toEqual(undefined);
   });
 });
 
 describe('getModalTableItems', () => {
   it('should get available items for current entity, then return them as javascript collection', () => {
-    expect(getModalTableItems(initialState)).toEqual([]);
+    expect(getModalTableItems(initialState, 'mockDependentEntity')).toEqual([]);
+  });
+  it('should get available items for current entity when entityName equals reasonLists, then return them as javascript collection', () => {
+    expect(
+      getModalTableItems(
+        fromJS({
+          Entities: {
+            mockEntity: {
+              data: [
+                {
+                  id: '0000',
+                  name: 'mockName',
+                  active: true,
+                  reasonLists: ['0001']
+                }
+              ]
+            },
+            reasonLists: {
+              data: [
+                {
+                  id: '0001',
+                  name: 'mockName',
+                  active: true
+                },
+                {
+                  id: '0002',
+                  name: 'mockNam2',
+                  active: true
+                }
+              ]
+            }
+          }
+        }),
+        'reasonLists'
+      )
+    ).toEqual([{ active: true, id: '0001', name: 'mockName' }, { active: true, id: '0002', name: 'mockNam2' }]);
+  });
+  it('should get empty array since there are no data for dependentEntity', () => {
+    expect(getModalTableItems(initialState, 'mockDependentEntity2')).toEqual([]);
+  });
+  it('should get empty array since dependentEntity data is undefined', () => {
+    expect(
+      getModalTableItems(
+        fromJS({
+          Entities: {
+            mockEntity: {
+              data: [
+                {
+                  id: '0000',
+                  name: 'mockName',
+                  active: true,
+                  mockDependentEntity: ['0001']
+                }
+              ]
+            },
+            mockDependentEntity: {
+              data: undefined
+            }
+          }
+        }),
+        'mockDependentEntity'
+      )
+    ).toEqual([]);
   });
 });
 
 describe('getListSize', () => {
   it('should get size of collection of dependant members of current entity', () => {
-    expect(getListSize(initialState)).toEqual(0);
+    expect(getListSize(initialState)).toEqual(1);
   });
 });
 
 describe('availableItemsForList', () => {
   it('should get all available items to add to list, then return them as javascript collection', () => {
-    expect(availableItemsForList(initialState)).toEqual([]);
+    expect(
+      availableItemsForList(
+        fromJS({
+          Entities: {
+            mockEntity: {
+              data: [
+                {
+                  id: '0000',
+                  name: 'mockName',
+                  active: true,
+                  mockDependentEntity: ['0001']
+                }
+              ]
+            },
+            mockDependentEntity: {
+              data: [
+                {
+                  id: '0001',
+                  name: 'mockName',
+                  active: true
+                },
+                {
+                  id: '0002',
+                  name: 'mockNam2',
+                  active: true
+                }
+              ]
+            }
+          }
+        })
+      )
+    ).toEqual([{ active: true, id: '0002', name: 'mockNam2' }]);
+  });
+  it('should get empty array since dependentEntity has no entries in state', () => {
+    expect(
+      availableItemsForList(
+        fromJS({
+          Entities: {
+            mockEntity: {
+              data: [
+                {
+                  id: '0000',
+                  mockDependentEntity: ['0001']
+                }
+              ]
+            },
+            mockDependentEntity: {
+              data: undefined
+            }
+          }
+        })
+      )
+    ).toEqual([]);
   });
 });
 
 describe('getEntityListMembers', () => {
   it('should get all members from an entity, then return them as javascript collection', () => {
-    expect(getEntityListMembers(initialState)).toEqual([]);
+    expect(getEntityListMembers(initialState)).toEqual([{ active: true, id: '0001', name: 'mockName' }]);
+  });
+  it('should get empty array if members are undefined for selected entity', () => {
+    expect(
+      getEntityListMembers(
+        fromJS({
+          Entities: {
+            mockEntity: {
+              data: [
+                {
+                  id: '0000',
+                  mockDependentEntity: undefined
+                }
+              ]
+            },
+            mockDependentEntity: {
+              data: undefined
+            }
+          }
+        })
+      )
+    ).toEqual([]);
   });
 });
 
@@ -144,7 +309,59 @@ describe('selectedEntityIndex', () => {
 
 describe('availableEntityMembersForList', () => {
   it('should get members available for current entity list, then return them as javascript collection', () => {
-    expect(availableEntityMembersForList(initialState)).toEqual([]);
+    expect(
+      availableEntityMembersForList(
+        fromJS({
+          Entities: {
+            mockEntity: {
+              data: [
+                {
+                  id: '0000',
+                  name: 'mockName',
+                  active: true,
+                  mockDependentEntity: ['0001']
+                }
+              ]
+            },
+            mockDependentEntity: {
+              data: [
+                {
+                  id: '0001',
+                  name: 'mockName',
+                  active: true
+                },
+                {
+                  id: '0002',
+                  name: 'mockNam2',
+                  active: true
+                }
+              ]
+            }
+          }
+        })
+      )
+    ).toEqual([{ active: true, id: '0002', name: 'mockNam2' }]);
+  });
+  it('should get empty array since dependentEntity has no entries in state', () => {
+    expect(
+      availableEntityMembersForList(
+        fromJS({
+          Entities: {
+            mockEntity: {
+              data: [
+                {
+                  id: '0000',
+                  mockDependentEntity: ['0001']
+                }
+              ]
+            },
+            mockDependentEntity: {
+              data: undefined
+            }
+          }
+        })
+      )
+    ).toEqual([]);
   });
 });
 
@@ -156,6 +373,6 @@ describe('entityRemovedFromList', () => {
 
 describe('entityAddedToList', () => {
   it('should add member with the id, then return new list as javascript collection', () => {
-    expect(entityAddedToList(initialState, '0002')).toEqual(['0002']);
+    expect(entityAddedToList(initialState, '0002')).toEqual(['0002', '0001']);
   });
 });